Further documentation and refactoring
